fix(applications): allow immediate action click after input changes
- make action buttons clickable immediately after form input changes
- apply behavior across attack, defence, ability, characteristic, rune, and spirit dialogs
- centralize interaction handling in shared interactive roll application base class
