+++ src/actors/rqgActor.ts
@@ -21,11 +21,11 @@ import { AbilitySuccessLevelEnum } from "../rolls/AbilityRoll/AbilityRoll.defs";
 import type { CharacteristicRollOptions } from "../rolls/CharacteristicRoll/CharacteristicRoll.types";
 import { CharacteristicRoll } from "../rolls/CharacteristicRoll/CharacteristicRoll";
 import type { Characteristic, Characteristics } from "../data-model/actor-data/characteristics";
-import { CharacteristicRollDialogV2 } from "../applications/CharacteristicRollDialog/characteristicRollDialogV2";
+import { CharacteristicRollDialogV2 } from "../applications/characteristic-roll-dialog/characteristic-roll-dialog-v2";
 import type { AbilityRollOptions } from "../rolls/AbilityRoll/AbilityRoll.types";
-import { AbilityRollDialogV2 } from "../applications/AbilityRollDialog/abilityRollDialogV2";
+import { AbilityRollDialogV2 } from "../applications/ability-roll-dialog/ability-roll-dialog-v2";
 import { AbilityRoll } from "../rolls/AbilityRoll/AbilityRoll";
-import type { PartialAbilityItem } from "../applications/AbilityRollDialog/AbilityRollDialogData.types.ts";
+import type { PartialAbilityItem } from "../applications/ability-roll-dialog/ability-roll-dialog-data.types.ts";
 import type { ActorHealthState } from "../data-model/actor-data/attributes";
 import type { DamageType } from "@item-model/weaponDataModel.ts";
 import { dodgeBaseChance, jumpBaseChance } from "../items/skill-item/skillFormulas";
