--- src/applications/app-parts/rollFooter.types.ts
@@ -1,4 +1,0 @@
-export type RollFooterData = {
-  totalChance: number;
-  rollMode: foundry.dice.Roll.Mode; // read in onSubmit by checking the active class
-};
