--- src/applications/app-parts/rollModes.hbs
@@ -1,34 +1,0 @@
-<div data-roll-mode-parent class="split-button flex-none">
-  <button type="button"
-          class="ui-control icon fa-solid fa-globe"
-          data-action="rollMode"
-          data-roll-mode="publicroll"
-          aria-pressed="{{#if (eq rollMode "publicroll")}}true{{else}}false{{/if}}"
-          data-tooltip="{{localize "CHAT.RollPublic"}}"
-          aria-label="Public Roll">
-  </button>
-  <button type="button"
-          class="ui-control icon fa-solid fa-user-secret"
-          data-action="rollMode"
-          data-roll-mode="gmroll"
-          aria-pressed="{{#if (eq rollMode "gmroll")}}true{{else}}false{{/if}}"
-          data-tooltip="{{localize "CHAT.RollPrivate"}}"
-          aria-label="Private GM Roll">
-  </button>
-  <button type="button"
-          class="ui-control icon fa-solid fa-eye-slash"
-          data-action="rollMode"
-          data-roll-mode="blindroll"
-          aria-pressed="{{#if (eq rollMode "blindroll")}}true{{else}}false{{/if}}"
-          data-tooltip="{{localize "CHAT.RollBlind"}}"
-          aria-label="Blind GM Roll">
-  </button>
-  <button type="button"
-          class="ui-control icon fa-solid fa-user"
-          data-action="rollMode"
-          data-roll-mode="selfroll"
-          aria-pressed="{{#if (eq rollMode "selfroll")}}true{{else}}false{{/if}}"
-          data-tooltip="{{localize "CHAT.RollSelf"}}"
-          aria-label="Self Roll">
-  </button>
-</div>
