+++ src/applications/attack-flow/attack-dialog-v2.hbs
