+++ src/applications/attack-flow/attack-footer.types.ts
