+++ src/applications/attack-flow/combat-roll-header.hbs
