+++ src/applications/attack-flow/defence-dialog-v2.hbs
