+++ src/applications/attack-flow/defence-footer-data.types.ts
