+++ src/applications/attack-flow/defence-footer.hbs
@@ -2,7 +2,7 @@
   <div class="flexrow flex-wrap gap1rem fullwidth">
     <div class="target-box" data-tooltip="{{localize "RQG.Dialog.Common.TargetChance"}}">
       <i class="fa-solid fa-bullseye"></i><br>
-      {{totalChance}}%
+      <span data-total-chance>{{totalChance}}%</span>
     </div>
 
     {{#if isSelectedParryWeaponBroken}}
