+++ src/applications/characteristic-roll-dialog/characteristic-roll-dialog-v2.hbs
