+++ src/applications/rune-magic-roll-dialog/rune-magic-roll-dialog-data.types.ts
@@ -1,5 +1,5 @@
 import type { RollHeaderData } from "../app-parts/rollHeader.types.ts";
-import type { RollFooterData } from "../app-parts/rollFooter.types.ts";
+import type { RollFooterData } from "../app-parts/roll-footer.types.ts";
 
 /** Minimal rune data needed for the roll rune dialog to avoid complex type recursion */
 export type PartialRuneItem = {
