+++ src/applications/rune-magic-roll-dialog/rune-magic-roll-dialog-v2.ts
@@ -1,12 +1,6 @@
 import { systemId } from "../../system/config";
 import { templatePaths } from "../../system/loadHandlebarsTemplates";
-import {
-  assertDocumentSubType,
-  getDomDataset,
-  getSpeakerFromItem,
-  localize,
-  RqgError,
-} from "../../system/util";
+import { assertDocumentSubType, getSpeakerFromItem, localize, RqgError } from "../../system/util";
 import type { RqgActor } from "@actors/rqgActor.ts";
 import { RqgItem } from "@items/rqgItem.ts";
 import {
@@ -18,21 +12,35 @@ import type { RuneMagicRollOptions } from "../../rolls/RuneMagicRoll/RuneMagicRo
 import type {
   RuneMagicRollDialogContext,
   RuneMagicRollDialogFormData,
-} from "./RuneMagicRollDialogData.types.ts";
-import type { PartialAbilityItem } from "../AbilityRollDialog/AbilityRollDialogData.types.ts";
+} from "./rune-magic-roll-dialog-data.types.ts";
+import type { PartialAbilityItem } from "../ability-roll-dialog/ability-roll-dialog-data.types.ts";
 import { ItemTypeEnum } from "@item-model/itemTypes.ts";
 import type { RuneMagicItem } from "@item-model/runeMagicDataModel.ts";
 import type { CultItem } from "@item-model/cultDataModel.ts";
-import type { DeepPartial } from "fvtt-types/utils";
-import { getDefaultRollMode, getSelectedRollMode } from "../app-parts/rollMode";
+import {
+  getConfiguredRollModeOptions,
+  getDefaultRollMode,
+  getSelectedRollMode,
+} from "../app-parts/roll-mode";
+import { RqgInteractiveRollApplicationBase } from "../app-parts/rqg-interactive-roll-application-base";
 
-const { ApplicationV2, HandlebarsApplicationMixin } = foundry.applications.api;
+export class RuneMagicRollDialogV2 extends RqgInteractiveRollApplicationBase {
+  protected override getLivePreviewFormBehaviorConfig() {
+    return {
+      submitButtonSelectorForBlurGuard: "button[data-ability-roll]",
+      updateLivePreview: () => this.updateLivePreview(),
+    };
+  }
 
-export class RuneMagicRollDialogV2 extends HandlebarsApplicationMixin(
-  ApplicationV2<RuneMagicRollDialogContext>,
-) {
-  override get element(): HTMLFormElement {
-    return super.element as HTMLFormElement;
+  private computeTotalChance(formData: RuneMagicRollDialogFormData): number {
+    const eligibleRunes = getEligibleRunes(this.spellItem);
+    const usedRune = eligibleRunes.find((r) => r.id === formData.usedRuneId);
+    return (
+      Number(usedRune?.system.chance ?? 0) +
+      Number(formData.augmentModifier ?? 0) +
+      Number(formData.meditateModifier ?? 0) +
+      Number(formData.otherModifier ?? 0)
+    );
   }
 
   private static augmentOptions: SelectOptionData<number>[] = [
@@ -71,7 +79,6 @@ export class RuneMagicRollDialogV2 extends HandlebarsApplicationMixin(
   ];
 
   private spellItem: RuneMagicItem;
-  private rollMode: foundry.dice.Roll.Mode;
 
   constructor(
     spellItem: RuneMagicItem,
@@ -79,7 +86,6 @@ export class RuneMagicRollDialogV2 extends HandlebarsApplicationMixin(
   ) {
     super(options);
     this.spellItem = spellItem;
-    this.rollMode = getDefaultRollMode();
   }
 
   static override DEFAULT_OPTIONS = {
@@ -159,32 +165,19 @@ export class RuneMagicRollDialogV2 extends HandlebarsApplicationMixin(
         Number(formData.meditateModifier ?? 0) +
         Number(formData.otherModifier ?? 0),
       rollMode: this.rollMode,
+      rollModes: getConfiguredRollModeOptions(),
     };
   }
 
-  override async _onRender(
-    context: DeepPartial<RuneMagicRollDialogContext>,
-    options: DeepPartial<foundry.applications.api.ApplicationV2.RenderOptions>,
-  ): Promise<void> {
-    super._onRender(context, options);
-    this.element
-      .querySelector<HTMLElement>("[data-roll-mode-parent]")
-      ?.addEventListener("click", this.onChangeRollMode.bind(this));
-  }
-
-  override _onChangeForm(): void {
-    this.render();
-  }
-
-  private onChangeRollMode(event: MouseEvent): void {
-    const target = event.target as HTMLButtonElement;
-    const newRollMode = getSelectedRollMode(getDomDataset(target, "roll-mode"));
-    if (!newRollMode) {
-      return; // Clicked outside the buttons, or not a valid roll mode
+  private updateLivePreview(): void {
+    const totalChanceElement = this.element.querySelector<HTMLElement>("[data-total-chance]");
+    if (!totalChanceElement) {
+      return;
     }
-    this.rollMode = newRollMode;
 
-    this.render();
+    const formData = new foundry.applications.ux.FormDataExtended(this.element, {})
+      .object as RuneMagicRollDialogFormData;
+    totalChanceElement.textContent = `${this.computeTotalChance(formData)}%`;
   }
 
   private static async onSubmit(
