+++ src/applications/spirit-magic-roll-dialog/spirit-magic-roll-dialog-v2.hbs
